NTP-1164: Restructure the code folders to align with the enquiry routes
