+++ Application/Common/Models/Enquiry/Manage/EnquirerViewResponseModel.cs
@@ -1,0 +1,3 @@
+namespace Application.Common.Models.Enquiry.Manage;
+
+public record EnquirerViewResponseModel : EnquiryResponseBaseModel;
