+++ Application/Common/Models/Enquiry/Manage/EnquirerViewTuitionPartnerDetailsModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Application.Common.Models.Enquiry.Respond;
+﻿namespace Application.Common.Models.Enquiry.Manage;
 
 public record EnquirerViewTuitionPartnerDetailsModel
 {
