--- Application/Common/Models/Enquiry/Respond/EnquirerViewResponseModel.cs
@@ -1,3 +1,0 @@
-namespace Application.Common.Models.Enquiry.Respond;
-
-public record EnquirerViewResponseModel : EnquiryResponseModel;
