+++ Application/Common/Models/Enquiry/Respond/ViewAndCaptureEnquiryResponseModel.cs
@@ -1,6 +1,6 @@
 namespace Application.Common.Models.Enquiry.Respond;
 
-public record ViewAndCaptureEnquiryResponseModel : EnquiryResponseModel
+public record ViewAndCaptureEnquiryResponseModel : EnquiryResponseBaseModel
 {
     public string EnquiryResponseCloseDateFormatted { get; set; } = string.Empty;
 }
