+++ Infrastructure/Repositories/TuitionPartnerEnquiryRepository.cs
@@ -1,5 +1,5 @@
 using Application.Common.Interfaces.Repositories;
-using Application.Common.Models.Enquiry.Respond;
+using Application.Common.Models.Enquiry.Manage;
 using Application.Extensions;
 using Domain;
 using Microsoft.EntityFrameworkCore;
