+++ UI/Pages/Enquiry/Manage/AllEnquirerResponses.cshtml
@@ -1,5 +1,5 @@
 @page "/enquiry/{supportReferenceNumber}"
-@model UI.Pages.Enquiry.Respond.AllEnquirerResponses
+@model UI.Pages.Enquiry.Manage.AllEnquirerResponses
 @using GovUk.Frontend.AspNetCore.TagHelpers
 @{
   ViewData["Title"] = "View all responses";
