+++ UI/Pages/Enquiry/Manage/EnquirerResponse.cshtml
@@ -1,5 +1,5 @@
 @page "/enquiry/{supportReferenceNumber}/{tuitionPartnerSeoUrl}"
-@model UI.Pages.Enquiry.Respond.EnquirerResponse
+@model UI.Pages.Enquiry.Manage.EnquirerResponse
 @{
   var baseServiceUrl = Request.GetBaseServiceUrl();
   var backLink = $"{baseServiceUrl}/enquiry/{Model.SupportReferenceNumber}?Token={Model.Data.Token}";
