+++ UI/Pages/Enquiry/Respond/CheckYourAnswers.cshtml.cs
@@ -1,5 +1,8 @@
+using Application.Commands.Enquiry.Respond;
 using Application.Common.Interfaces;
 using Application.Common.Models.Enquiry.Respond;
+using Application.Queries.Enquiry;
+using UI.Models;
 
 namespace UI.Pages.Enquiry.Respond;
